agents/community/langgraph-agents/pyproject.toml
[project]
name = "langgraph-agents"
version = "0.0.1"
description = "Lightweight web research and summarization assistant."
authors = [
    { name = "Lance Martin" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "beeai-sdk==0.0.16",
    "acp-sdk", # version determined by beeai-sdk
    "langgraph>=0.2.55",
    "langchain-community>=0.3.9",
    "tavily-python>=0.5.0",
    "langchain-ollama>=0.2.1",
    "duckduckgo-search>=7.3.0",
    "beautifulsoup4>=4.13.3",
    "langchain-openai>=0.3.7",
    "openinference-instrumentation-langchain>=0.1.36",
]

[project.scripts]
langgraph-agents = "langgraph_agents.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
